feat: ロールダイスロジック実装
