+++ lib/dice_roller.dart
@@ -1,0 +1,50 @@
+import 'dart:math';
+import 'package:flutter/material.dart';
+
+final randomizer = Ramdom();
+
+class DiceRoller extends StatefulWidget {
+  const DiceRoller({super.key});
+
+  @override
+  State<DiceRoller> createState() {
+    return _DiceRollerState();
+  }
+}
+
+class _DiceRollerState extends State<DiceRoller> {
+  var currentDiceRoll = 1;
+
+  void rollDice() {
+    setState(() {
+      currentDiceRoll = randomizer.nextInt(6) + 1;
+    });
+  }
+
+  @override
+  Widget build(context) {
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        Image.asset(
+          'assets/images/dice-$currentDiceRoll.png';,
+          width: 200,
+        ),
+        const SizedBox(height: 20),
+        TextButton(
+          onPressed: rollDice,
+          style: TextButton.styleFrom(
+            // padding: const EdgeInsets.only(
+            //   top: 20,
+            // ),
+            foregroundColor: Colors.white,
+            textStyle: const TextStyle(
+              fontSize: 28,
+            ),
+          ),
+          child: const Text('Roll Dice'),
+        )
+      ],
+    );
+  }
+}
